% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stats.R
\name{print.stats}
\alias{print.stats}
\title{Generic S3 print method for objects of class stats}
\usage{
\method{print}{stats}(x, ...)
}
\arguments{
\item{x}{The object to be printed.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Generic S3 print method for objects of class stats
}
