% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_chisq.R
\name{sta_chisq}
\alias{sta_chisq}
\title{Pearson's or McNemar's Chi-squared Test for Count Data}
\usage{
sta_chisq(
  x = NULL,
  y = NULL,
  paired = FALSE,
  obj = TRUE,
  correct = TRUE,
  p = rep(1/length(x), length(x)),
  rescale.p = FALSE,
  simulate.p.value = FALSE,
  B = 2000
)
}
\arguments{
\item{x}{a numeric vector or matrix. x and y can also both be factors.}

\item{y}{a numeric vector; ignored if x is a matrix. If x is a factor, y should be a factor of the same length.}

\item{paired}{a logical variable indicating whether to treat the two variances as being equal.}

\item{obj}{a logical whether the test object of stats::chisq.test() or stats::mcnemar.test should be returned.}

\item{correct}{a logical indicating whether to apply continuity correction when computing the test statistic for 2 by 2 tables: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves. No correction is done if simulate.p.value = TRUE.}

\item{p}{a vector of probabilities of the same length of x. An error is given if any entry of p is negative.}

\item{rescale.p}{a logical scalar; if TRUE then p is rescaled (if necessary) to sum to 1. If rescale.p is FALSE, and p does not sum to 1, an error is given.}

\item{simulate.p.value}{a logical indicating whether to compute p-values by Monte Carlo simulation.}

\item{B}{an integer specifying the number of replicates used in the Monte Carlo test.}
}
\description{
Performs chi-squared contingency table tests and goodness-of-fit tests.
}
\examples{
\dontrun{
library(magrittr)
data(big_five, package = "inductive")

kont <- table(big_five$Geschlecht, big_five$Extraversion > 4)

sta_chisq(kont)

kont \%>\% 
  sta_chisq(paired = TRUE)

}

}
