% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_var.R
\name{sta_var}
\alias{sta_var}
\title{F Test to Compare Two Variances}
\usage{
sta_var(
  data = NULL,
  formula = NULL,
  x = NULL,
  y = NULL,
  ratio = 1,
  alternative = "two.sided",
  conf.level = 0.95,
  obj = TRUE,
  na.action = getOption("na.action"),
  ...
)
}
\arguments{
\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs a factor with two levels giving the corresponding groups.}

\item{x}{numeric vectors of data values, or fitted linear model objects (inheriting from class "lm").}

\item{y}{numeric vectors of data values, or fitted linear model objects (inheriting from class "lm").}

\item{ratio}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{obj}{a logical whether the test object of stats::var.test() should be returned.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Performs an F test to compare the variances of two samples from normal populations.
}
\examples{
\dontrun{
library(magrittr)
data(big_five, package = "inductive")

big_five \%>\% 
  sta_var(Extraversion ~ Gender)

big_five \%>\% 
  sta_var(Extraversion ~ Neuroticism)

big_five \%>\% 
  sta_var(x = Extraversion, y = Neuroticism)

sta_var(big_five, x = Extraversion, y = Neuroticism)

sta_var(x = big_five$Extraversion, y = big_five$Neuroticism)

sta_var(big_five, Extraversion ~ Gender)
}

}
