% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_wilcox.R
\name{sta_wilcox}
\alias{sta_wilcox}
\title{Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
sta_wilcox(data = NULL, formula = NULL, x = NULL, y = NULL,
  alternative = c("two.sided", "less", "greater"), mu = 0,
  paired = FALSE, exact = NULL, correct = TRUE, conf.int = FALSE,
  conf.level = 0.95, obj = TRUE, na.action = getOption("na.action"),
  ...)
}
\arguments{
\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs a factor with two levels giving the corresponding groups.}

\item{x}{numeric vector of data values. Non-finite (e.g., infinite or missing) values will be omitted.}

\item{y}{an optional numeric vector of data values: as with x non-finite values will be omitted.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{mu}{a number specifying an optional parameter used to form the null hypothesis. See ‘Details’.}

\item{paired}{a logical indicating whether you want a paired test.}

\item{exact}{a logical indicating whether an exact p-value should be computed.}

\item{correct}{a logical indicating whether to apply continuity correction in the normal approximation for the p-value.}

\item{conf.int}{a logical indicating whether a confidence interval should be computed.}

\item{conf.level}{confidence level of the interval.}

\item{obj}{a logical whether the test object of stats::t.test() should be returned.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Performs one- and two-sample Wilcoxon tests on vectors of data; the latter is also known as ‘Mann-Whitney’ test.
}
\examples{
\dontrun{
library(magrittr)
data(big_five, package = "inductive")

res <- big_five \%>\%
  sta_wilcox(Extraversion ~ Geschlecht)

res$result
res$obj

big_five \%>\%
  sta_wilcox(Extraversion ~ Neurotizismus)

big_five \%>\%
  sta_wilcox(x = Extraversion, y = Neurotizismus)

big_five \%>\%
  sta_wilcox(x = Neurotizismus, y = Extraversion)
}

}
